[onert] Change SubgraphIndex value type
It changes type for value from uint32 to uint16.

tflite and circle schema define Subgraphs as following:
```
  subgraphs:[SubGraph];
```
It can have 2^32 Subgraphs.
But onert assumes the number of subgraph does not exceed 2^16.
We will remove the assumption when we encounter such case.

ONE-DCO-1.0-Signed-off-by: Sanggyu Lee <sg5.lee@samsung.com>

+++ runtime/onert/core/include/ir/Index.h
@@ -36,7 +36,7 @@ struct IOIndexTag;
 using IOIndex = ::onert::util::Index<uint32_t, IOIndexTag>;
 
 struct SubgraphIndexTag;
-using SubgraphIndex = ::onert::util::Index<uint32_t, SubgraphIndexTag>;
+using SubgraphIndex = ::onert::util::Index<uint16_t, SubgraphIndexTag>;
 
 struct ModelIndexTag;
 using ModelIndex = ::onert::util::Index<uint16_t, ModelIndexTag>;
